+++ CHANGELOG.md
@@ -7,9 +7,18 @@ Adheres to [Semantic Versioning](http://semver.org/).
 ## 1.1.0 (TBD)
 
 * coordinate-reference-systems-java version 1.1.0
-* Additional CRS Parser Swiss Oblique Mercator support
-* CRS Parser SCALE_FACTOR_ON_PSEUDO_STANDARD_PARALLEL support
-* CRS Parser CO_LATITUDE_OF_CONE_AXIS support
+* CRS Parser fixes and improvements
+  * Swiss Oblique Mercator projections
+  * Bessel vs Bessel Namibia Ellipsoid parsing
+  * Lambert Cylindrical Equal Area (Spherical) operation method
+  * Polar Stereographic (variants A,B, & C) operation methods
+  * Equidistant Cylindrical (Spherical) operation method
+  * Mercator (variants A & B) operation methods
+  * Popular Visualisation Pseudo Mercator operation method
+  * Lambert Conic Conformal (1SP & 2SP) operation methods
+  * Hotine Oblique Mercator (variant B) operation method for Swiss Oblique Mercator
+  * scale factor on pseudo standard parallel operation parameter
+  * co-latitude of cone axis operation parameter
 * CRS Definition to PROJ parameters conversion and projection creation
 
 ## [1.0.0](https://github.com/ngageoint/projections-java/releases/tag/1.0.0) (06-24-2021)